Add ports state test
